Task_2/mlcubes/model/project/requirements.txt
PyYAML
typer
numpy
SimpleITK